requirements.txt
google-generativeai
typing-extensions
python-dotenv
moviepy
pydub
streamlit
tenacity
pydantic
groq
opencv-python
